Carrito en JS con fetch
Ahora se puede:
- Eliminar productos del carrito y afecta tanto a la base de datos como a la vista.  El producto deja de verse en la pagina.
- Modificar la cantidad de items en el carrito. Estos cambios modifican el subtotal de dicho producto. 

Ambos tipos de cambios afectan al total de la compra y el total de items en carrito que se ve en la bolsa de la esquina.

+++ app/Http/Controllers/CarritoController.php
@@ -66,16 +66,23 @@ class CarritoController extends Controller{
     public function apiCantidad(Request $req){
         $user_id = $req->user_id;
         $producto_id = $req->producto_id;
-        Carrito::where('producto_id', $producto_id)->where('user_id', $user_id)->where('estado','0')->first();
-        $existe->cantidad += $req ->cantidad;
-        $existe->save();
+        $carrito = Carrito::where('producto_id', $req->producto_id)->where('user_id', $user_id)->where('estado','0')->first();
+        $carrito->cantidad = $req->cantidad;
+        $carrito->save();
+
         return $this->api($user_id);
     }
 
     public function apiBorrar(Request $req){
         $user_id = $req->user_id;
-        $carrito = Carrito::where('id', $req->id)->where('user_id', $user_id)->where('estado', '0')->first();
-        $carrito->delete();
+        $producto_id = $req->producto_id;
+        $carrito = Carrito::where('producto_id', $producto_id)->where('user_id', $user_id)->where('estado', '0')->first();
+        if($carrito){
+            $carrito->delete();
+        } else {
+            $carrito2 = Carrito::find($req->item_id);
+            $carrito2->delete;
+        }
         return $this->api($user_id);
     }
